Fill in report endpoint
