+++ src/modules/site-v2/templates/tools/phenotype_database/report.html
@@ -8,16 +8,16 @@
             <h2 class="h3 text-center mb-5">Trait Information</h2>
             <h3 class="h4 text-dark">Trait 1</h3>
             <p><strong>Name:</strong></p>
-            <p>Short Name</p>
+            <p>{{ report['trait_1']['trait_name'] }}</p>
             <p><strong>Description</strong></p>
-            <p>The short description of the trait</p>
+            <p>{{ report['trait_1']['description_short'] }}</p>
             <h3 class="h4 text-dark">Trait 2</h3>
             <p><strong>Name:</strong></p>
-            <p>Short Name</p>
+            <p>{{ report['trait_2']['trait_name'] }}</p>
             <p><strong>Description</strong></p>
-            <p>The short description of the trait</p>
+            <p>{{ report['trait_2']['description_short'] }}</p>
             <h3 class="h4 text-dark">Correlation</h3>
-            <p>correlation value</p>
+            <p>{{ result.correlation|round(4) }}</p>
         </div>
         <!-- /Input Card -->
         <!-- Tool column -->
